remove virtual authenticator path for selenium
