+++ test/e2e/page-objects/flows/onboarding.flow.ts
@@ -51,8 +51,10 @@ export const handleSidepanelPostOnboarding = async (
   );
 
   // Navigate directly to home page in current window
-  // With sidepanel enabled, this ensures we load home page in the test window
-  await driver.driver.get(`${driver.extensionUrl}/home.html`);
+  // With sidepanel enabled, this ensures we load home page in the test window.
+  // Uses the driver-level method (not the raw `driver.driver` session) so the
+  // flow works on both the Selenium and Playwright paths.
+  await driver.openNewURL(`${driver.extensionUrl}/home.html`);
 
   // Wait for the home page to fully load
   const headerNavbar = new HeaderNavbar(driver);
